full_app/device/requirements.txt
gpiozero
luma.core
luma.oled
pillow
RPi.GPIO